Conflicts:
	src/integrations.F90
	src/libnegf.F90

+++ src/mat_def.F90
@@ -1,9 +1,9 @@
 !!--------------------------------------------------------------------------!
 !! libNEGF: a general library for Non-Equilibrium Green's functions.        !
 !! Copyright (C) 2012                                                       !
-!!                                                                          ! 
+!!                                                                          !
 !! This file is part of libNEGF: a library for                              !
-!! Non Equilibrium Green's Function calculation                             ! 
+!! Non Equilibrium Green's Function calculation                             !
 !!                                                                          !
 !! Developers: Alessandro Pecchia, Gabriele Penazzi                         !
 !! Former Conctributors: Luca Latessa, Aldo Di Carlo                        !
@@ -15,12 +15,12 @@
 !!                                                                          !
 !!  You should have received a copy of the GNU Lesser General Public        !
 !!  License along with libNEGF.  If not, see                                !
-!!  <http://www.gnu.org/licenses/>.                                         !  
+!!  <http://www.gnu.org/licenses/>.                                         !
 !!--------------------------------------------------------------------------!
 
 
 Module mat_def
-  use ln_precision  
+  use ln_precision
   use ln_allocation
   implicit none
   private
@@ -108,7 +108,7 @@ interface print_mat
 end interface
 
 interface read_mat
-   module procedure zRead_CSR  
+   module procedure zRead_CSR
 end interface
 
 interface writemem
@@ -130,10 +130,10 @@ end interface
 
 Type z_CSR
   integer :: nnz = 0
-  integer :: nrow = 0  
+  integer :: nrow = 0
   integer :: ncol = 0
   logical :: sorted = .false.
-  complex(kind=dp), DIMENSION(:), ALLOCATABLE :: nzval  
+  complex(kind=dp), DIMENSION(:), ALLOCATABLE :: nzval
   integer, DIMENSION(:), ALLOCATABLE :: colind
   integer, DIMENSION(:), ALLOCATABLE :: rowpnt
 end Type z_CSR
@@ -142,12 +142,12 @@ end Type z_CSR
 
 Type z_CSC
   integer :: nnz = 0
-  integer :: nrow = 0  
+  integer :: nrow = 0
   integer :: ncol = 0
   logical :: sorted = .false.
-  complex(kind=dp), DIMENSION(:), ALLOCATABLE :: nzval 
-  integer, DIMENSION(:), ALLOCATABLE :: rowind 
-  integer, DIMENSION(:), ALLOCATABLE :: colpnt 
+  complex(kind=dp), DIMENSION(:), ALLOCATABLE :: nzval
+  integer, DIMENSION(:), ALLOCATABLE :: rowind
+  integer, DIMENSION(:), ALLOCATABLE :: colpnt
 end Type z_CSC
 
 !MSR Complex Structure definition (Modified Sparse Row Format)
@@ -157,61 +157,61 @@ Type z_MSR
   integer :: nrow = 0
   integer :: ncol = 0
   logical :: sorted = .false.
-  complex(kind=dp), DIMENSION(:), ALLOCATABLE :: nzval 
-  integer, DIMENSION(:), ALLOCATABLE :: index 
+  complex(kind=dp), DIMENSION(:), ALLOCATABLE :: nzval
+  integer, DIMENSION(:), ALLOCATABLE :: index
 end Type z_MSR
 
 Type z_COO
   integer :: nnz = 0
   integer :: nrow = 0
   integer :: ncol = 0
-  integer, DIMENSION(:), ALLOCATABLE :: index_i 
-  integer, DIMENSION(:), ALLOCATABLE :: index_j 
-  complex(kind=dp), DIMENSION(:), ALLOCATABLE :: nzval  
+  integer, DIMENSION(:), ALLOCATABLE :: index_i
+  integer, DIMENSION(:), ALLOCATABLE :: index_j
+  complex(kind=dp), DIMENSION(:), ALLOCATABLE :: nzval
 end Type z_COO
 
 Type z_EXT_COO
   integer :: nnz = 0
   integer :: nrow = 0
   integer :: ncol = 0
-  integer, DIMENSION(:), ALLOCATABLE :: index_i 
-  integer, DIMENSION(:), ALLOCATABLE :: index_j 
+  integer, DIMENSION(:), ALLOCATABLE :: index_i
+  integer, DIMENSION(:), ALLOCATABLE :: index_j
   complex(kind=dp), DIMENSION(:), ALLOCATABLE :: nzval
   logical, DIMENSION(:), ALLOCATABLE :: first
-end Type z_EXT_COO 
+end Type z_EXT_COO
 
 Type z_DNS
   integer :: nrow = 0
   integer :: ncol = 0
-  complex(kind=dp), DIMENSION(:,:), ALLOCATABLE :: val  
-end Type z_DNS 
+  complex(kind=dp), DIMENSION(:,:), ALLOCATABLE :: val
+end Type z_DNS
 
 Type r_DNS3
   integer :: nrow = 0
   integer :: ncol = 0
   integer :: npoints = 0
-  real(kind=dp), DIMENSION(:,:,:), ALLOCATABLE :: val  
+  real(kind=dp), DIMENSION(:,:,:), ALLOCATABLE :: val
 end Type r_DNS3
 
 Type z_DNS3
   integer :: nrow = 0
   integer :: ncol = 0
   integer :: npoints = 0
-  complex(kind=dp), DIMENSION(:,:,:), ALLOCATABLE :: val  
+  complex(kind=dp), DIMENSION(:,:,:), ALLOCATABLE :: val
 end Type z_DNS3
 
 Type c_DNS3
   integer :: nrow = 0
   integer :: ncol = 0
   integer :: npoints = 0
-  complex(kind=sp), DIMENSION(:,:,:), ALLOCATABLE :: val  
+  complex(kind=sp), DIMENSION(:,:,:), ALLOCATABLE :: val
 end Type c_DNS3
- 
+
 
 Type z_vec
    integer :: len = 0
-   complex(dp), dimension(:), allocatable :: val 
-   integer, dimension(:), allocatable :: ind 
+   complex(dp), dimension(:), allocatable :: val
+   integer, dimension(:), allocatable :: ind
 end Type z_vec
 
 Type z_RGM
@@ -220,24 +220,24 @@ Type z_RGM
 end Type z_RGM
 
 
- 
+
 !CSR Complex Structure definition (Compressed Sparse Row Format)
 Type r_CSR
   integer :: nnz = 0
   integer :: nrow = 0
   integer :: ncol = 0
   logical :: sorted = .false.
-  real(kind=dp), DIMENSION(:), ALLOCATABLE :: nzval  
-  integer, DIMENSION(:), ALLOCATABLE :: colind 
-  integer, DIMENSION(:), ALLOCATABLE :: rowpnt 
+  real(kind=dp), DIMENSION(:), ALLOCATABLE :: nzval
+  integer, DIMENSION(:), ALLOCATABLE :: colind
+  integer, DIMENSION(:), ALLOCATABLE :: rowpnt
 end Type r_CSR
 
 Type s_CSR
   integer :: nnz = 0
-  integer :: nrow = 0  
+  integer :: nrow = 0
   integer :: ncol = 0
   logical :: sorted = .false.
-  real(kind=sp), DIMENSION(:), ALLOCATABLE :: nzval  
+  real(kind=sp), DIMENSION(:), ALLOCATABLE :: nzval
   integer, DIMENSION(:), ALLOCATABLE :: colind
   integer, DIMENSION(:), ALLOCATABLE :: rowpnt
 end Type s_CSR
@@ -246,12 +246,12 @@ end Type s_CSR
 
 Type r_CSC
   integer :: nnz = 0
-  integer :: nrow = 0 
+  integer :: nrow = 0
   integer :: ncol = 0
   logical :: sorted = .false.
-  real(kind=dp), DIMENSION(:), ALLOCATABLE :: nzval 
-  integer, DIMENSION(:), ALLOCATABLE :: rowind 
-  integer, DIMENSION(:), ALLOCATABLE :: colpnt 
+  real(kind=dp), DIMENSION(:), ALLOCATABLE :: nzval
+  integer, DIMENSION(:), ALLOCATABLE :: rowind
+  integer, DIMENSION(:), ALLOCATABLE :: colpnt
 end Type r_CSC
 
 !MSR Real Structure definition (Modified Sparse Row Format)
@@ -261,31 +261,31 @@ Type r_MSR
   integer :: ncol = 0
   integer :: nrow = 0
   logical :: sorted = .false.
-  real(kind=dp), DIMENSION(:), ALLOCATABLE :: nzval 
-  integer, DIMENSION(:), ALLOCATABLE :: index 
+  real(kind=dp), DIMENSION(:), ALLOCATABLE :: nzval
+  integer, DIMENSION(:), ALLOCATABLE :: index
 end Type r_MSR
 
 Type r_COO
   integer :: nnz = 0
   integer :: nrow = 0
   integer :: ncol = 0
-  integer, DIMENSION(:), ALLOCATABLE :: index_i 
-  integer, DIMENSION(:), ALLOCATABLE :: index_j 
-  real(kind=dp), DIMENSION(:), ALLOCATABLE :: nzval  
-end Type r_COO 
+  integer, DIMENSION(:), ALLOCATABLE :: index_i
+  integer, DIMENSION(:), ALLOCATABLE :: index_j
+  real(kind=dp), DIMENSION(:), ALLOCATABLE :: nzval
+end Type r_COO
 
 Type r_DNS
-  integer :: nrow = 0 
+  integer :: nrow = 0
   integer :: ncol = 0
-  real(kind=dp), DIMENSION(:,:), ALLOCATABLE :: val  
-end Type r_DNS 
+  real(kind=dp), DIMENSION(:,:), ALLOCATABLE :: val
+end Type r_DNS
 
 
 Type c_DNS
-  integer :: nrow = 0 
+  integer :: nrow = 0
   integer :: ncol = 0
-  complex(kind=sp), DIMENSION(:,:), ALLOCATABLE :: val  
-end Type c_DNS 
+  complex(kind=sp), DIMENSION(:,:), ALLOCATABLE :: val
+end Type c_DNS
 ! *******************************************************************
 contains
 
@@ -294,7 +294,7 @@ subroutine zcreate_pCSR(mat,nrow,ncol,nnz)
   type(z_CSR), pointer :: mat
   integer :: nrow, ncol, nnz
   integer :: ierr
-  
+
   allocate(mat,stat=ierr)
   if(ierr.ne.0) stop 'ERROR: pointer zCSR not allocated'
 
@@ -309,7 +309,7 @@ subroutine zcreate_pCSR(mat,nrow,ncol,nnz)
   endif
 
   call log_allocate(mat%rowpnt,nrow+1)
-    
+
 end subroutine zcreate_pCSR
 ! ------------------------------------------------------------------
 
@@ -328,7 +328,7 @@ subroutine zcreate_CSR(mat,nrow,ncol,nnz)
   endif
 
   call log_allocate(mat%rowpnt,nrow+1)
-    
+
 end subroutine zcreate_CSR
 ! ------------------------------------------------------------------
 
@@ -337,7 +337,7 @@ subroutine zinit_CSR(mat)
 
   integer :: k
 
-  if(mat%nnz.ne.mat%nrow) STOP 'cannot initialize matrix (nnz != nrow)' 
+  if(mat%nnz.ne.mat%nrow) STOP 'cannot initialize matrix (nnz != nrow)'
 
   do k=1,Mat%nrow
      Mat%nzval(k)=0.d0
@@ -345,7 +345,7 @@ subroutine zinit_CSR(mat)
      Mat%rowpnt(k)=k
   enddo
   Mat%rowpnt(mat%nrow+1)=mat%nrow+1
-  
+
 end subroutine zinit_CSR
 ! ------------------------------------------------------------------
 subroutine zcreate_id_CSR(mat,nrow)
@@ -372,7 +372,7 @@ subroutine zclone_id_CSR(mat, matH)
   call zcreate_CSR(mat,matH%nrow,matH%nrow,matH%nnz)
 
   nrow = matH%nrow
-  nnz = matH%nnz 
+  nnz = matH%nnz
 
   do i=1,nrow+1
      mat%rowpnt(i)=matH%rowpnt(i)
@@ -381,10 +381,10 @@ subroutine zclone_id_CSR(mat, matH)
   do i=1,nnz
      mat%colind(i)=matH%colind(i)
   enddo
-  
+
   do i=1,nrow
-    do j =mat%rowpnt(i), mat%rowpnt(i+1)-1  
-      if (mat%colind(j) .eq. i) then 
+    do j =mat%rowpnt(i), mat%rowpnt(i+1)-1
+      if (mat%colind(j) .eq. i) then
          mat%nzval(j)=(1.d0,0.d0)
       else
          mat%nzval(j)=(0.d0,0.d0)
@@ -427,7 +427,7 @@ subroutine zrecreate_CSR(mat)
 
   call zdestroy_CSR(mat)
   call zcreate_CSR(mat,ncol,ncol,ncol)
-    
+
    do k=1,ncol
      mat%nzval(k)=0.d0
      mat%colind(k)=k
@@ -441,7 +441,7 @@ subroutine zdestroy_CSR(mat1,mat2,mat3,mat4,mat5,mat6,mat7,mat8)
 
   type(z_CSR) :: mat1
   type(z_CSR), optional :: mat2,mat3,mat4,mat5,mat6,mat7,mat8
-  
+
   mat1%nnz=0
   mat1%nrow=0
   mat1%ncol=0
@@ -576,12 +576,12 @@ subroutine zPrint_CSR(id,sp,fmt)
   type(z_CSR) :: sp
   integer :: id,i
   logical :: fmt
-  
+
   if(fmt) then
      write(id,*) 'Nrow: ',sp%nrow
-     write(id,*) 'Ncol: ',sp%ncol 
+     write(id,*) 'Ncol: ',sp%ncol
      write(id,*) 'Nnz: ',sp%nnz
-     
+
      write(id,*) 'Nzval array'
      do i=1,sp%nnz
         write(id,*) sp%nzval(i)
@@ -592,15 +592,15 @@ subroutine zPrint_CSR(id,sp,fmt)
      enddo
      write(id,*) 'Rowpnt array'
      if(sp%nrow.gt.0) then
-     do i=1,sp%nrow+1   
+     do i=1,sp%nrow+1
         write(id,*) sp%rowpnt(i)
      enddo
      endif
   else
      write(id) sp%nrow
-     write(id) sp%ncol 
+     write(id) sp%ncol
      write(id) sp%nnz
-     
+
      !write(id) 'Nzval array'
      do i=1,sp%nnz
         write(id) sp%nzval(i)
@@ -628,9 +628,9 @@ subroutine zRead_CSR(id,sp,fmt)
 
   if(fmt) then
      read(id,*) tmpst,sp%nrow
-     read(id,*) tmpst,sp%ncol 
+     read(id,*) tmpst,sp%ncol
      read(id,*) tmpst,sp%nnz
-     
+
      read(id,*) tmpst
      do i=1,sp%nnz
         read(id,*) sp%nzval(i)
@@ -640,7 +640,7 @@ subroutine zRead_CSR(id,sp,fmt)
         read(id,*) sp%colind(i)
      enddo
      read(id,*) tmpst
-     
+
      if(sp%nrow.gt.0) then
      do i=1,sp%nrow+1
         read(id,*) sp%rowpnt(i)
@@ -648,17 +648,17 @@ subroutine zRead_CSR(id,sp,fmt)
      endif
   else
      read(id) sp%nrow
-     read(id) sp%ncol 
+     read(id) sp%ncol
      read(id) sp%nnz
-     
+
      do i=1,sp%nnz
         read(id) sp%nzval(i)
      enddo
-     
+
      do i=1,sp%nnz
         read(id) sp%colind(i)
      enddo
-     
+
      if(sp%nrow.gt.0) then
      do i=1,sp%nrow+1
         read(id) sp%rowpnt(i)
@@ -695,7 +695,7 @@ subroutine zdestroy_CSC(mat1,mat2,mat3,mat4,mat5,mat6,mat7,mat8)
 
   type(z_CSC) :: mat1
   type(z_CSC), optional :: mat2,mat3,mat4,mat5,mat6,mat7,mat8
-  
+
   mat1%nnz=0
   mat1%nrow=0
   mat1%ncol=0
@@ -805,10 +805,10 @@ subroutine zPrint_CSC(sp)  !Identical to Print_CSR
 
   type(z_CSC) :: sp
   integer :: i
-  
+
   do i=1,sp%nnz
     write(*,*) sp%nzval(i)
-  enddo  
+  enddo
 
 end subroutine zprint_CSC
 ! ------------------------------------------------------------------
@@ -834,12 +834,12 @@ end subroutine zcreate_MSR
 subroutine zdestroy_MSR(mat)
 
   type(z_MSR) :: mat
-  
+
   mat%nnz=0
   mat%nrow=0
   mat%ncol=0
 
-  if (allocated(mat%nzval)) then 
+  if (allocated(mat%nzval)) then
      call log_deallocate(mat%nzval)
      call log_deallocate(mat%index)
   endif
@@ -851,10 +851,10 @@ subroutine zprint_MSR(sp)
 
   type(z_MSR) :: sp
   integer :: i
-  
+
   do i=1,sp%nnz
     write(*,*) sp%nzval(i)
-  enddo  
+  enddo
 
 end subroutine zprint_MSR
 ! ------------------------------------------------------------------
@@ -883,7 +883,7 @@ subroutine zdestroy_COO(mat1,mat2,mat3,mat4,mat5,mat6,mat7,mat8)
 
   type(z_COO) :: mat1
   type(z_COO), optional :: mat2,mat3,mat4,mat5,mat6,mat7,mat8
-  
+
   mat1%nnz=0
   mat1%nrow=0
   mat1%ncol=0
@@ -994,23 +994,23 @@ subroutine zprint_COO(id,sp,fmt)
   type(z_COO) :: sp
   integer :: id,i
   logical :: fmt
-  
+
   if(fmt) then
      write(id,*) sp%nrow,'Nrow'
      write(id,*) sp%ncol,'Ncol'
      write(id,*) sp%nnz,'Nnz'
-  
-     do i=1,sp%nnz  
+
+     do i=1,sp%nnz
         write(id,*) sp%index_i(i),sp%index_j(i),sp%nzval(i)
      enddo
 
   else
 
      write(id) sp%nrow
-     write(id) sp%ncol 
+     write(id) sp%ncol
      write(id) sp%nnz
-  
-     do i=1,sp%nnz  
+
+     do i=1,sp%nnz
         write(id) sp%index_i(i),sp%index_j(i),sp%nzval(i)
      enddo
 
@@ -1041,18 +1041,18 @@ subroutine zcreate_EXT_COO(mat,nrow,ncol,nnz)
   endif
 
   mat%first(:) = .false.
-  
+
 end subroutine zcreate_EXT_COO
 ! ------------------------------------------------------------------
 
 subroutine zdestroy_EXT_COO(mat)
 
   type(z_EXT_COO) :: mat
-  
+
   mat%nnz=0
   mat%nrow=0
   mat%ncol=0
-  if (allocated(mat%nzval)) then 
+  if (allocated(mat%nzval)) then
      call log_deallocate(mat%nzval)
      call log_deallocate(mat%index_i)
      call log_deallocate(mat%index_j)
@@ -1166,7 +1166,7 @@ subroutine zdestroy_DNS(mat1,mat2,mat3,mat4,mat5,mat6,mat7,mat8)
 
   type(z_DNS) :: mat1
   type(z_DNS), optional :: mat2,mat3,mat4,mat5,mat6,mat7,mat8
-  
+
   mat1%nrow=0
   mat1%ncol=0
 
@@ -1251,7 +1251,7 @@ subroutine cdestroy_DNS(mat1,mat2,mat3,mat4,mat5,mat6,mat7,mat8)
 
   type(c_DNS) :: mat1
   type(c_DNS), optional :: mat2,mat3,mat4,mat5,mat6,mat7,mat8
-  
+
   mat1%nrow=0
   mat1%ncol=0
 
@@ -1369,7 +1369,7 @@ subroutine rrecreate_CSR(mat)
 
   call rdestroy_CSR(mat)
   call rcreate_CSR(mat,ncol,ncol,ncol)
-    
+
    do k=1,ncol
      mat%nzval(k)=0.d0
      mat%colind(k)=k
@@ -1384,7 +1384,7 @@ subroutine rdestroy_CSR(mat1,mat2,mat3,mat4,mat5,mat6,mat7,mat8)
 
   type(r_CSR) :: mat1
   type(r_CSR), optional :: mat2,mat3,mat4,mat5,mat6,mat7,mat8
-  
+
   mat1%nnz=0
   mat1%nrow=0
   mat1%ncol=0
@@ -1495,22 +1495,22 @@ subroutine rPrint_CSR(id,sp)
 
   type(r_CSR) :: sp
   integer :: i,id
-  
+
   write(id,*) 'Nrow is ',sp%nrow
   write(id,*) 'Nnz is ',sp%nnz
 
   write(id,*) 'Nzval array'
   do i=1,sp%nnz
     write(id,*) sp%nzval(i)
-  enddo 
+  enddo
   write(id,*) 'Colind array'
   do i=1,sp%nnz
     write(id,*) sp%colind(i)
-  enddo  
+  enddo
   write(id,*) 'Rowpnt array'
   do i=1,sp%nrow+1
     write(id,*) sp%rowpnt(i)
-  enddo 
+  enddo
 
 end subroutine rPrint_CSR
 ! ------------------------------------------------------------------
@@ -1520,9 +1520,9 @@ subroutine rcreate_CSC(mat,nrow,ncol,nnz)
 
   type(r_CSC) :: mat
   integer :: nrow, ncol, nnz
-  
+
   if(nrow.eq.0.or.ncol.eq.0) STOP 'ERROR: (rcreate_CSC) nrow or ncol = 0'
-   
+
   mat%nnz=nnz
   mat%nrow=nrow
   mat%ncol=ncol
@@ -1538,14 +1538,14 @@ end subroutine rcreate_CSC
 subroutine rdestroy_CSC(mat)
 
   type(r_CSC) :: mat
-  
+
   mat%nnz=0
   mat%nrow=0
   mat%ncol=0
-  if (allocated(mat%nzval)) then 
+  if (allocated(mat%nzval)) then
      call log_deallocate(mat%nzval)
      call log_deallocate(mat%rowind)
-  endif  
+  endif
   call log_deallocate(mat%colpnt)
 
 end subroutine rdestroy_CSC
@@ -1555,10 +1555,10 @@ subroutine rPrint_CSC(sp)  !Identical to Print_CSR
 
   type(r_CSC) :: sp
   integer :: i
-  
+
   do i=1,sp%nnz
     write(*,*) sp%nzval(i)
-  enddo  
+  enddo
 
 end subroutine rprint_CSC
 ! ------------------------------------------------------------------
@@ -1584,10 +1584,10 @@ end subroutine rcreate_MSR
 subroutine rdestroy_MSR(mat)
 
   type(r_MSR) :: mat
-  
+
   mat%nnz=0
   mat%nrow=0
-  if (allocated(mat%nzval)) then 
+  if (allocated(mat%nzval)) then
      call log_deallocate(mat%nzval)
      call log_deallocate(mat%index)
   endif
@@ -1598,10 +1598,10 @@ subroutine rprint_MSR(sp)
 
   type(r_MSR) :: sp
   integer :: i
-  
+
   do i=1,sp%nnz
     write(*,*) sp%nzval(i)
-  enddo  
+  enddo
 
 end subroutine rprint_MSR
 ! ------------------------------------------------------------------
@@ -1629,7 +1629,7 @@ subroutine rdestroy_COO(mat1,mat2,mat3,mat4,mat5,mat6,mat7,mat8)
 
   type(r_COO) :: mat1
   type(r_COO), optional :: mat2,mat3,mat4,mat5,mat6,mat7,mat8
-  
+
   mat1%nnz=0
   mat1%nrow=0
   mat1%ncol=0
@@ -1740,7 +1740,7 @@ subroutine rcreate_DNS(mat,nrow,ncol)
 
  type(r_DNS) :: mat
  integer :: nrow, ncol
-  
+
 
   if(nrow.eq.0.or.ncol.eq.0) STOP 'ERROR: (rcreate_DNS) nrow or ncol = 0'
 
@@ -1752,10 +1752,10 @@ end subroutine rcreate_DNS
 ! ------------------------------------------------------------------
 
 subroutine rdestroy_DNS(mat1,mat2,mat3,mat4,mat5,mat6,mat7,mat8)
-  
+
   type(r_DNS) :: mat1
   type(r_DNS), optional :: mat2,mat3,mat4,mat5,mat6,mat7,mat8
-  
+
   mat1%nrow=0
   mat1%ncol=0
 
@@ -1840,10 +1840,10 @@ subroutine rprint_COO(id,sp)
 
   type(r_COO) :: sp
   integer :: id,i
-  
+
   do i=1,sp%nnz
     write(id,*) sp%index_i(i),sp%index_j(i),sp%nzval(i)
-  enddo  
+  enddo
 
 end subroutine rprint_COO
 ! ------------------------------------------------------------------
@@ -1852,7 +1852,7 @@ subroutine zwriteMem_CSR(id,mat)
 
   type(z_CSR) :: mat
   integer :: id,dec
-  integer(8) :: mem
+  integer(long) :: mem
   character(3) :: str
 
   mem=mat%nnz*(2*dp+4)+(mat%nrow+2)*4
